refactor: Moved encoding to util functions
