+++ src/utils/index.ts
@@ -11,3 +11,4 @@ export * from './permissions';
 export * from './proposals';
 export * from './scheme';
 export * from './transaction';
+export * from './encodingCalls';
